requirements.txt
numpy>=1.18.2
pytest-cov
